fix(quota): humanize stale-data freshness label
The quota panel formatted `now - updated_at` as raw seconds, so a
recent claude refresh rendered as `138ago` and an older codex pull
showed as `341493ago` (also visually clipped to `341493ag` in narrow
columns). The neighboring mcp panel already had a `fmt_age` helper
that buckets into s/m/h/d, but it was a private function on `mcp.rs`.

Promote `fmt_age` to `ui::fmt_age` next to `fmt_mem_kb` / `fmt_tokens`
so both panels share it, switch the quota label to use it, and drop
the now-unused `quota.ago` locale key (en + zh). The mcp panel
behavior is unchanged.
