+++ src/locale.rs
@@ -200,7 +200,6 @@ static LOCALE_EN: LazyLock<std::collections::HashMap<&str, &str>> = LazyLock::ne
     m.insert("quota.run_codex", "run codex once");
     m.insert("quota.total", "total");
     m.insert("quota.now", "now");
-    m.insert("quota.ago", "ago");
 
     // Projects panel
     m.insert("projects.no_git", "no git");
@@ -442,7 +441,6 @@ static LOCALE_ZH: LazyLock<std::collections::HashMap<&str, &str>> = LazyLock::ne
     m.insert("quota.run_codex", "运行一次 codex");
     m.insert("quota.total", "总计");
     m.insert("quota.now", "现在");
-    m.insert("quota.ago", "前");
 
     // Projects panel
     m.insert("projects.no_git", "非 Git");
